git-svn-id: https://cscl.googlecode.com/svn/trunk@128 157b09d4-0768-11df-9bc7-f7babce8d823

+++ CSCL/Games/Manasource/Monster.cs
@@ -44,7 +44,7 @@ namespace CSCL.Games.Manasource
 		{
 			get
 			{
-				return Attributes.HP*(Attributes.AttackMin+(Attributes.AttackMin+Attributes.AttackDelta));
+				return (Int64)Attributes.HP*(Attributes.AttackMin+(Attributes.AttackMin+Attributes.AttackDelta));
 			}
 		}
